tests/wpt/tests/webnn/permissions-policy/webnn-disabled-by-permissions-policy.https.html

<!DOCTYPE html>
<body>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=/permissions-policy/resources/permissions-policy.js></script>
<script src=/common/get-host-info.sub.js></script>
<script>
'use strict';

const same_origin_src =
    '/webnn/permissions-policy/permissions-policy-webnn.html';
const cross_origin_src =
    get_host_info().HTTPS_REMOTE_ORIGIN + same_origin_src;
const header = 'Permissions policy header "webnn=()"';
const attribute = 'allow="webnn" attribute';

promise_test(async t => {
  await promise_rejects_dom(t, 'SecurityError', navigator.ml.createContext());
}, `${header} disallows the top-level document.`);

async_test(t => {
  test_feature_availability('navigator.ml.createContext()', t, same_origin_src,
      expect_feature_unavailable_default);
}, `${header} disallows same-origin iframes.`);

async_test(t => {
  test_feature_availability('navigator.ml.createContext()', t, cross_origin_src,
      expect_feature_unavailable_default);
}, `${header} disallows cross-origin iframes.`);

async_test(t => {
  test_feature_availability('navigator.ml.createContext()', t, same_origin_src,
      expect_feature_unavailable_default, 'webnn');
}, `${header} disallows same-origin iframes despite the ${attribute}.`);

async_test(t => {
  test_feature_availability('navigator.ml.createContext()', t, cross_origin_src,
      expect_feature_unavailable_default, 'webnn');
}, `${header} disallows cross-origin iframes despite the ${attribute}.`);

</script>
</body>
